+++ src/components/TodoList/TodoList.tsx
@@ -7,37 +7,24 @@ import { Status } from '../../types/enums/Status';
 export const TodoList: React.FC = () => {
   const { todos, filter } = useContext(StateContext);
 
-  const todoItems: JSX.Element[] = [];
-
-  todos.forEach((todo) => {
-    switch (filter) {
-      case Status.All:
-        todoItems.push(<TodoItem key={todo.id} todoItem={todo} />);
-        break;
-
-      case Status.Active:
-        if (!todo.completed) {
-          todoItems.push(<TodoItem key={todo.id} todoItem={todo} />);
-        }
-
-        break;
-
-      case Status.Completed:
-        if (todo.completed) {
-          todoItems.push(<TodoItem key={todo.id} todoItem={todo} />);
-        }
-
-        break;
-
-      default:
-        todoItems.push(<TodoItem key={todo.id} todoItem={todo} />);
-        break;
-    }
-  });
-
   return (
-    <ul className="todo-list" data-cy="todoList">
-      {todoItems}
+    <ul className="todo-list" data-cy="todosList">
+      {
+        todos
+          .filter((todo) => {
+            switch (filter) {
+              case Status.Active:
+                return !todo.completed;
+
+              case Status.Completed:
+                return todo.completed;
+
+              default:
+                return true;
+            }
+          })
+          .map((todo) => <TodoItem key={todo.id} todoItem={todo} />)
+      }
     </ul>
   );
 };
